+++ Kernel/Output/HTML/HeaderMeta/AgentTicketSearch.pm
@@ -32,12 +32,6 @@ sub Run {
     # get layout object
     my $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
 
-    my $Session = '';
-    if ( !$LayoutObject->{SessionIDCookie} ) {
-        $Session = ';' . $LayoutObject->{SessionName} . '='
-            . $LayoutObject->{SessionID};
-    }
-
     # get config object
     my $ConfigObject = $Kernel::OM->Get('Kernel::Config');
 
@@ -50,7 +44,7 @@ sub Run {
             Type  => 'application/opensearchdescription+xml',
             Title => $Title,
             Href  => $LayoutObject->{Baselink} . 'Action=' . $Param{Config}->{Action}
-                . ';Subaction=OpenSearchDescriptionTicketNumber' . $Session,
+                . ';Subaction=OpenSearchDescriptionTicketNumber',
         },
     );
 
@@ -64,7 +58,7 @@ sub Run {
             Type  => 'application/opensearchdescription+xml',
             Title => $Title,
             Href  => $LayoutObject->{Baselink} . 'Action=' . $Param{Config}->{Action}
-                . ';Subaction=OpenSearchDescriptionFulltext' . $Session,
+                . ';Subaction=OpenSearchDescriptionFulltext',
         },
     );
     return 1;