+++ var/httpd/htdocs/js/test/Core.App.UnitTest.js
@@ -16,20 +16,10 @@ Core.App = (function (Namespace) {
         QUnit.module('Core.App');
 
         QUnit.test('Core.App.GetSessionInformation()', function(Assert){
-            Assert.expect(2);
+            Assert.expect(1);
 
-            Core.Config.Set('SessionName', 'CSID');
-            Core.Config.Set('SessionID', '1234');
-            Core.Config.Set('CustomerPanelSessionName', 'CPanelSID');
             Core.Config.Set('ChallengeToken', 'C123');
 
-            Assert.deepEqual(Core.App.GetSessionInformation(), {
-                CSID: '1234',
-                CPanelSID: '1234',
-                ChallengeToken: 'C123'
-            });
-
-            Core.Config.Set('SessionIDCookie', true);
             Assert.deepEqual(Core.App.GetSessionInformation(), {
                 ChallengeToken: 'C123'
             });